Move methods with content type definitions
Move methods which return MIME content types:
- #multipart_signed_content_type
- #multipart_encrypted_content_type
- #encrypted_message_content_type
to RFC1847 helper, as these content types are all defined in the related
RFC.

Move methods:
- #sign_protocol
- #encryption_protocol
- #message_integrity_algorithm
- #encryption_control_information
to RFC3156 helper as they all define things common to all e-mails secured
with Open PGP.

Methods have been made protected as there is no good reason to make them
a part of the public API.

In tests, some methods had to be stubbed in order to get tests back
green.
