+++ lib/enmail/adapters/gpgme.rb
@@ -81,42 +81,6 @@ module EnMail
       def build_crypto
         ::GPGME::Crypto.new(armor: true)
       end
-
-      public
-
-      def multipart_signed_content_type
-        protocol = sign_protocol
-        micalg = message_integrity_algorithm
-        %[multipart/signed; protocol="#{protocol}"; micalg="#{micalg}"]
-      end
-
-      def multipart_encrypted_content_type
-        protocol = encryption_protocol
-        %[multipart/encrypted; protocol="#{protocol}"]
-      end
-
-      # The encrypted message must have content type +application/octet-stream+,
-      # as defined in RFC 1847 p. 6.
-      def encrypted_message_content_type
-        "application/octet-stream"
-      end
-
-      def sign_protocol
-        "application/pgp-signature"
-      end
-
-      def encryption_protocol
-        "application/pgp-encrypted"
-      end
-
-      def message_integrity_algorithm
-        "pgp-sha1"
-      end
-
-      # As defined in RFC 3156
-      def encryption_control_information
-        "Version: 1"
-      end
     end
   end
 end
