+++ lib/enmail/helpers/rfc3156.rb
@@ -6,6 +6,24 @@ module EnMail
     #
     # See: https://tools.ietf.org/html/rfc3156
     module RFC3156
+      protected
+
+      def sign_protocol
+        "application/pgp-signature"
+      end
+
+      def encryption_protocol
+        "application/pgp-encrypted"
+      end
+
+      def message_integrity_algorithm
+        "pgp-sha1"
+      end
+
+      # As defined in RFC 3156
+      def encryption_control_information
+        "Version: 1"
+      end
     end
   end
 end
