+++ spec/unit/adapters/gpgme_spec.rb
@@ -173,48 +173,4 @@ RSpec.describe EnMail::Adapters::GPGME do
       expect(retval).to be_a_pgp_encrypted_message.encrypted_for(*recipients)
     end
   end
-
-  describe "#multipart_signed_content_type" do
-    subject { adapter.method(:multipart_signed_content_type) }
-
-    it "returns a string" do
-      expect(subject.call).to be_a(String)
-    end
-
-    it "has a MIME type multipart/signed" do
-      retval_segments = subject.call.split(/\s*;\s*/)
-      expect(retval_segments[0]).to eq("multipart/signed")
-    end
-
-    it "tells about SHA1 message integrity algorithm" do
-      retval_segments = subject.call.split(/\s*;\s*/)
-      micalg_def = %[micalg="pgp-sha1"]
-      expect(retval_segments[1..-1]).to include(micalg_def)
-    end
-
-    it "tells about PGP protocol" do
-      retval_segments = subject.call.split(/\s*;\s*/)
-      protocol_def = %[protocol="application/pgp-signature"]
-      expect(retval_segments[1..-1]).to include(protocol_def)
-    end
-  end
-
-  describe "#multipart_encrypted_content_type" do
-    subject { adapter.method(:multipart_encrypted_content_type) }
-
-    it "returns a string" do
-      expect(subject.call).to be_a(String)
-    end
-
-    it "has a MIME type multipart/encrypted" do
-      retval_segments = subject.call.split(/\s*;\s*/)
-      expect(retval_segments[0]).to eq("multipart/encrypted")
-    end
-
-    it "tells about PGP protocol" do
-      retval_segments = subject.call.split(/\s*;\s*/)
-      protocol_def = %[protocol="application/pgp-encrypted"]
-      expect(retval_segments[1..-1]).to include(protocol_def)
-    end
-  end
 end
